backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==23.1
djoser==2.1.0
Pillow==9.3.0
requests==2.26.0
python-dotenv==1.0.1
drf-extra-fields==3.7.0
